1.01

+++ Assets/Scripts/Events/AddMoneyEvent.cs
@@ -6,6 +6,6 @@ using UnityEngine.Events;
 /// <summary>
 /// Add money event
 /// </summary>
-public class AddMoneyEvent : UnityEvent<int>
+public class AddMoneyEvent : UnityEvent<float>
 {
 }
