+++ Assets/Scripts/Events/SavePlayerMoneyEvent.cs
@@ -6,6 +6,6 @@ using UnityEngine.Events;
 /// <summary>
 /// Saves player money
 /// </summary>
-public class SavePlayerMoneyEvent : UnityEvent<int>
+public class SavePlayerMoneyEvent : UnityEvent<float>
 {
 }
